pythonWebTemplate/requirements.txt
beautifulsoup4==4.5.3
Flask==0.12
Werkzeug==0.16
requests
flask-cors
python-dotenv
mysql-connector-python